server/requirements.txt
flask 
flask-sqlalchemy
flask-migrate
python-docx
mysqlclient